pyproject.toml
[tool.poetry]
name = "poketwo"
version = "0.1.0"
description = "A small Pokécord clone"
authors = ["Oliver Ni <oliver.ni@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.0"
umongo = "3.0.0b10"
motor = "^2.2.0"
jishaku = "^1.19.1"
dnspython = "^2.0.0"
suntime = "^1.2.5"
geocoder = "^1.38.1"
humanfriendly = "^8.2"
unidecode = "^1.1.1"
discord-flags = "^2.1.1"
quart = "^0.13.1"
durations_nlp = "^1.0.1"
aioredis = "^1.3.1"
uvloop = "^0.14.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus.git"}
aioredis-lock = "^0.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100